Separated Lobby and Game controllers
Made the lobby an SSR'd view that contains the form to join a game

The Game Controller is responsable for taking a valid player name and
saving them to the client's session

Also added HanabiWeb.LiveMonitor to add unmount to the game_live view.
So when the process ends by moving away from the page the player exits
the lobby.